requirements.txt
joblib~=0.16.0
imblearn~=0.0
scikit-learn~=0.23.1
Flask-SQLAlchemy~=2.4.1
SQLAlchemy~=1.3.15
PyYAML~=5.4
Flask~=1.1.1
pytest~=5.4.2
pandas~=1.1.5
numpy~=1.21.6
botocore~= 1.15.32
boto3~=1.12.32
s3fs~=0.5.1
fsspec~=0.8.4
pymysql~=1.0.2
markupsafe==2.0.1